% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calofood-package.R
\docType{package}
\name{Calofood-package}
\alias{Calofood}
\alias{Calofood-package}
\title{Calofood: R Package to Perform Numeric Functions, Plot Visuals and Launch the Cal-o-food App}
\description{
"Calofood" is a comprehensive tool designed to access and analyze nutritional information from well-known fast food restaurants. This package offers a user-friendly interface to access calorie counts and other nutritional details for a diverse selection of fast food menu items which will cater to your specefic health goal. With Calofood, effortlessly retrieve and compare nutritional information for different items, aiding you in making well-informed dietary choices when eating out. Whether your goal is to manage calorie intake, monitor specific dietary preferences, or opt for healthier fast food options, Calofood equips you with the knowledge necessary to support your nutritional objectives.
}
\author{
\strong{Maintainer}: Disha Rathod \email{drat0009@student.monash.edu}

}
\keyword{internal}
