% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericfunction.R
\name{comp_int}
\alias{comp_int}
\title{Calculates Compound Interest}
\usage{
comp_int(principal, rate, time)
}
\arguments{
\item{principal}{Initial investment amount (positive numeric value).}

\item{rate}{Annual interest rate as a decimal (positive numeric value).}

\item{time}{Duration in years for interest calculation (positive numeric value).}
}
\value{
The compunded interest amount of the input.
}
\description{
This function accepts these three inputs, computes the compound interest,
and provides the result. Additionally, it incorporates a mechanism for
error checking to guarantee that all inputs are positive values since
negative values or zeros would be illogical within the context of compound interest
computations.
Principal (numeric): The initial sum of money or investment,
referred to as the principal amount, used for computing compound interest.
It must be a positive numeric value, symbolizing the initial monetary amount.
Rate (numeric): The annual interest rate, expressed as a decimal,
governing the rate at which the principal amount increases over time due to interest.
It should also be a positive numeric value.
Time (numeric): The duration, in years, over which interest is computed.
This parameter signifies the period during which the principal amount accrues
interest. It must be a positive numeric value indicating the number of years.
}
\examples{
comp_int(450000,6,8)

}
