% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{foodplot}
\alias{foodplot}
\title{Bar Plot for A Vitamin Rich Diet}
\usage{
foodplot(resto)
}
\arguments{
\item{resto}{The name of the restaurant/restaurants}
}
\value{
A bar plot of the containing the food items for the restaurant entered
}
\description{
This function helps you to decide the food item that is best suited for you to have
a vitamin-rich diet. The input for this function can be one of the leading fast-food
joints as follows:
\itemize{
\item Mcdonalds
\item Chick Fil-A
\item Sonic
\item Arbys
\item Burger King
\item Dairy Queen
\item Subway
\item Taco Bell
\item All Restaurants (this is when you can't decide which restaurant you want to
eat at and want a combined response from all the restaurants listed above)
}
}
\examples{
foodplot("Mcdonalds")
foodplot("All Restaurants")

}
