% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{The Shiny App}
\usage{
run_app()
}
\value{
Shiny App
}
\description{
Praises you or someone with a random word.
This function launches the Shiny app called 'Cal-o-food,' where users can select
their health goals using nutrient information from a wide array of variables
in the dataset. It allows users to choose from a range of trending fast food
restaurants and provides a list of recommendations tailored to their selection.
The recommendations include food items that best align with the user's preferences.
The results are visualized, providing a customized output based on the user's input.
}
